refactor and improve `memberOrganizationService` transaction handling (#2849)
